Minor tweaks

+++ src/r_things.c
@@ -1289,11 +1289,11 @@ static void R_SortVisSprites(void)
 {
     if (num_vissprite)
     {
-        static size_t   num_vissprite_ptrs;
+        static unsigned int num_vissprite_ptrs;
 
-        if (num_vissprite_ptrs < (size_t)num_vissprite * 2)
+        if (num_vissprite_ptrs < num_vissprite * 2)
             vissprite_ptrs = I_Realloc(vissprite_ptrs,
-                (num_vissprite_ptrs = (size_t)num_vissprite_alloc * 2) * sizeof(*vissprite_ptrs));
+                (num_vissprite_ptrs = num_vissprite_alloc * 2) * sizeof(*vissprite_ptrs));
 
         for (int i = num_vissprite - 1; i >= 0; i--)
             vissprite_ptrs[i] = vissprites + i;
